requirements.txt
# Core Dependencies for Image Generation System
# PyTorch - Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Diffusion Models
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Model Management
huggingface-hub>=0.19.0

# Evaluation Metrics
scikit-image>=0.21.0
scipy>=1.11.0

# UI (Optional but recommended)
gradio>=4.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.7.0

# ControlNet specific
controlnet-aux>=0.0.7

# CLIP for evaluation
ftfy>=6.1.0
regex>=2023.0.0

# For advanced features
xformers>=0.0.22  # Optional: for memory efficient attention (requires GPU)